feat: batch cold account loads in light client (#2341)
* feat: optimize address batch pipeline

* format

* feat: stabilize address batch pipeline

* feat: batch cold account loads in light client

* fix: harden load batching and mixed decompression

* Fix prover startup and decompression load flow

* cleanup: harden prover startup polling

* format

* format

* cleanup

* cleanup

* refactor: simplify batch data length validation and remove redundant proof height checks

* refactor: remove unused output_queue_index parameter from into_in_token_data methods

+++ prover/client/src/proof_types/batch_update/proof_inputs.rs
@@ -175,7 +175,7 @@ pub fn get_batch_update_inputs<const HEIGHT: usize>(
         index_bytes[28..].copy_from_slice(&(*index).to_be_bytes());
         let nullifier = Poseidon::hashv(&[leaf, &index_bytes, &tx_hashes[i]]).unwrap();
         let (root, changelog_entry) =
-            compute_root_from_merkle_proof(nullifier, &merkle_proof_array, *index)?;
+            compute_root_from_merkle_proof(nullifier, &merkle_proof_array, *index as usize)?;
         new_root = root;
         changelog.push(changelog_entry);
         circuit_merkle_proofs.push(
